requirement.txt
torchtyping
pytorch_custom_utils
beartype
x_transformers
local_attention
vector_quantize_pytorch
classifier_free_guidance_pytorch
torch_geometric
gateloop_transformer
ema_pytorch
trimesh
wandb
libigl
matplotlib
plyfile
optimum
transformers==4.38.2
accelerate
tensorboard
deepspeed